+++ src/games/arma2oa.rs
@@ -1,12 +1,12 @@
 use crate::GDResult;
 use crate::protocols::valve;
-use crate::protocols::valve::{game, SteamID};
+use crate::protocols::valve::{game, SteamApp};
 
 pub fn query(address: &str, port: Option<u16>) -> GDResult<game::Response> {
     let valve_response = valve::query(address, match port {
         None => 2304,
         Some(port) => port
-    }, SteamID::ARMA2OA.as_app(), None, None)?;
+    }, SteamApp::ARMA2OA.as_engine(), None, None)?;
 
     Ok(game::Response::new_from_valve_response(valve_response))
 }